Update godelw
It was a little unclear what the purpose of this file was for when reading through the README.  

You can start the server using:

./godelw dep
./godelw run example

inspecting the file: godelw to understand what the 1st command ( ./godelw dep) was doing and why it needed to download files was unclear.  This little comment will help people quickly appreciate what is happening.